Fix bug when 0-dimension

+++ tests/chainer_tests/functions_tests/math_tests/test_squared_difference.py
@@ -41,7 +41,7 @@ class TestSquaredDifference(unittest.TestCase):
     def check_backward(self, x1, x2, g_data):
         x_data = (x1, x2)
         gradient_check.check_backward(
-            functions.SquaredDifference(), x_data, g_data, dtype=self.dtype, atol=1e-2, rtol=1e-2)
+            functions.SquaredDifference(), x_data, g_data, dtype=numpy.float64, atol=1e-2, rtol=1e-2)
 
     @condition.retry(3)
     def test_backward_cpu(self):
